+++ app/javascript/vue/tasks/images/new_image/store/actions/actions.js
@@ -1,6 +1,7 @@
 import ActionNames from './actionNames'
 import applyAttibutions from './applyAttibutions'
 import applyDepictions from './applyDepictions'
+import applyTags from './applyTags'
 import applyPixelToCentimeter from './applyPixelToCentimeter'
 import resetStore from './resetStore'
 import removeImage from './removeImage'
@@ -8,9 +9,13 @@ import removeImage from './removeImage'
 const ActionFunctions = {
   [ActionNames.ApplyAttributions]: applyAttibutions,
   [ActionNames.ApplyDepictions]: applyDepictions,
+  [ActionNames.ApplyTags]: applyTags,
   [ActionNames.ApplyPixelToCentimeter]: applyPixelToCentimeter,
   [ActionNames.ResetStore]: resetStore,
   [ActionNames.RemoveImage]: removeImage
 }
 
-export { ActionNames, ActionFunctions }
+export {
+  ActionNames,
+  ActionFunctions
+}